requirements.txt
python-geoip==1.2
python-geoip-geolite2==2015.303
python-geoip-python3==1.3
